Fix smallPalindrom overflow issue.
The algorithm did not find the number "7036267126217626307" but numbers
after it, a small check showed that there is an overflow due to the
compiler casting a number to something smallet than ull, the casting
fix this issue.

+++ LargeNumbersFinder.h
@@ -75,7 +75,7 @@ void checkLargePalindromes(const unsigned long num) {
   palindromeToCheck = ((__uint128_t)num << (num_length + 1)) | reversedLowerHalf;
   printIfDecimalPalindromeLarge(palindromeToCheck);
 
-  palindromeToCheck = palindromeToCheck | (1 << num_length);
+  palindromeToCheck = palindromeToCheck | ((__uint128_t)1 << num_length);
   printIfDecimalPalindromeLarge(palindromeToCheck);
 }
